pyproject.toml
[project]
name = "apple-calendar-mcp"
version = "0.1.0"
description = "Apple Calendar MCP Server - EventKit 기반 캘린더 관리"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "Allen Kim"}]
keywords = ["apple-calendar", "eventkit", "mcp", "calendar"]

dependencies = [
    "mcp[cli]>=1.2.1",
    "pyobjc-framework-EventKit>=11.0",
]

[project.scripts]
apple-calendar-mcp = "apple_calendar_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apple_calendar_mcp"]
